Move cuda files
