+++ taichi/backends/cuda/CMakeLists.txt
@@ -1,0 +1,24 @@
+# ./taichi/backends/cuda/CMakeLists.txt
+
+set(CUDA_RHI cuda_rhi)
+add_library(${CUDA_RHI})
+target_sources(${CUDA_RHI}
+  PRIVATE
+    cuda_device.cpp
+    cuda_caching_allocator.cpp
+    cuda_context.cpp
+    cuda_driver.cpp
+    cuda_profiler.cpp
+    cupti_toolkit.cpp
+  )
+
+target_include_directories(${CUDA_RHI}
+  PRIVATE
+    ${PROJECT_SOURCE_DIR}
+    ${PROJECT_SOURCE_DIR}/external/eigen
+    ${PROJECT_SOURCE_DIR}/external/spdlog/include
+    ${PROJECT_SOURCE_DIR}/external/glfw/include
+    ${LLVM_INCLUDE_DIRS}
+  )
+
+# depends on cuda
