+++ taichi/codegen/cuda/codegen_cuda.cpp
@@ -1,4 +1,4 @@
-#include "taichi/backends/cuda/codegen_cuda.h"
+#include "taichi/codegen/cuda/codegen_cuda.h"
 
 #include <vector>
 #include <set>
