+++ taichi/runtime/cuda/aot_module_loader_impl.cpp
@@ -1,9 +1,9 @@
-#include "taichi/backends/cuda/aot_module_loader_impl.h"
+#include "taichi/runtime/cuda/aot_module_loader_impl.h"
 #include "taichi/runtime/llvm/llvm_aot_module_loader.h"
 
 #include "taichi/runtime/llvm/llvm_offline_cache.h"
 #include "taichi/runtime/program_impls/llvm/llvm_program.h"
-#include "taichi/backends/cuda/codegen_cuda.h"
+#include "taichi/codegen/cuda/codegen_cuda.h"
 
 namespace taichi {
 namespace lang {
