+++ taichi/runtime/llvm/CMakeLists.txt
@@ -32,4 +32,5 @@ endif()
 
 if (TI_WITH_CUDA)
     target_link_libraries(llvm_runtime PRIVATE ${llvm_ptx_libs})
+    target_link_libraries(llvm_runtime PRIVATE cuda_rhi)
 endif()
