+++ taichi/runtime/program_impls/llvm/CMakeLists.txt
@@ -15,3 +15,9 @@ target_include_directories(llvm_program_impl
   )
 
 target_link_libraries(llvm_program_impl PRIVATE llvm_codegen)
+
+if (TI_WITH_CUDA)
+    target_link_libraries(llvm_program_impl PRIVATE cuda_codegen)
+    target_link_libraries(llvm_program_impl PRIVATE cuda_runtime)
+    target_link_libraries(llvm_program_impl PRIVATE cuda_rhi)
+endif()
