+++ taichi/runtime/program_impls/llvm/llvm_program.cpp
@@ -9,8 +9,8 @@
 #include "taichi/backends/cpu/aot_module_builder_impl.h"
 
 #if defined(TI_WITH_CUDA)
-#include "taichi/backends/cuda/aot_module_builder_impl.h"
-#include "taichi/backends/cuda/codegen_cuda.h"
+#include "taichi/runtime/cuda/aot_module_builder_impl.h"
+#include "taichi/codegen/cuda/codegen_cuda.h"
 #endif
 
 namespace taichi {
